+++ src/49-ears.js
@@ -706,6 +706,7 @@ const ears = (() => {
     'arms.wide': 'arms out wide', 'arms.down': 'arms back down',
     'mouth.open': 'her mouth wide open', 'mouth.close': 'her mouth closed',
     'pet': 'you petting her hair',
+    'legs.spread': 'her legs apart', 'legs.close': 'her legs together',
     // Her hands up to the back of her head, and then the swap. See `tieHair`
     // in 43-jadrija.js: it is the same latch the console has always driven,
     // with the two seconds of gesture in front of it that it never had.